refactor(@svelte-put/copy): migration to vanilla JS
