+++ .changeset/sour-cycles-laugh.md
@@ -1,0 +1,5 @@
+---
+'@svelte-put/copy': major
+---
+
+Migrated to vanilla JS (tracked at #203)
