+++ packages/actions/copy/README.md
@@ -48,7 +48,6 @@ This package is part of the [@svelte-put][github.monorepo] family. For contribut
 [github.monorepo]: https://github.com/vnphanquang/svelte-put
 [github.changelog]: https://github.com/vnphanquang/svelte-put/blob/main/packages/actions/copy/CHANGELOG.md
 [github.issues]: https://github.com/vnphanquang/svelte-put/issues?q=
-[github.api]: https://github.com/vnphanquang/svelte-put/blob/main/packages/actions/copy/api/docs/index.md
 
 <!-- heading badge -->
