--- packages/actions/copy/api-extractor.json
@@ -1,6 +1,0 @@
-{
-  "$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json",
-  "extends": "@svelte-put/apirc/base.json",
-  "projectFolder": ".",
-  "mainEntryPointFilePath": "<projectFolder>/dist/src/index.d.ts"
-}
