--- packages/actions/copy/api/copy.api.md
@@ -1,45 +1,0 @@
-## API Report File for "@svelte-put/copy"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-
-import type { ActionReturn } from 'svelte/action';
-
-// @public
-export function copy<K extends keyof HTMLElementEventMap>(node: HTMLElement, parameters?: Partial<CopyParameters<K>>): ActionReturn<Partial<CopyParameters<K>>, CopyAttributes>;
-
-// @public
-export interface CopyAttributes {
-    // (undocumented)
-    'on:copied'?: (event: CustomEvent<CopyDetail>) => void;
-}
-
-// @public
-export interface CopyDetail {
-    text: string;
-}
-
-// @public
-export interface CopyParameters<K extends keyof HTMLElementEventMap> {
-    enabled: boolean;
-    event: K | K[];
-    synthetic: boolean;
-    text: string | TextResolver<K>;
-    trigger: HTMLElement;
-}
-
-// @public
-export function copyToClipboard(text: string): Promise<void> | undefined;
-
-// @public
-export type TextResolver<K extends keyof HTMLElementEventMap> = (input: TextResolverInput<K>) => string | Promise<string>;
-
-// @public
-export interface TextResolverInput<K extends keyof HTMLElementEventMap> {
-    event: HTMLElementEventMap[K];
-    node: HTMLElement;
-    trigger: HTMLElement;
-}
-
-```
