--- packages/actions/copy/api/docs/copy.copy.md
@@ -1,27 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@svelte-put/copy](./copy.md) &gt; [copy](./copy.copy.md)
-
-## copy() function
-
-Copy text to clipboard on `click` event
-
-**Signature:**
-
-```typescript
-export declare function copy<K extends keyof HTMLElementEventMap>(node: HTMLElement, parameters?: Partial<CopyParameters<K>>): ActionReturn<Partial<CopyParameters<K>>, CopyAttributes>;
-```
-
-## Parameters
-
-|  Parameter | Type | Description |
-|  --- | --- | --- |
-|  node | HTMLElement | HTMLElement to register action |
-|  parameters | Partial&lt;[CopyParameters](./copy.copyparameters.md)<!-- -->&lt;K&gt;&gt; | _(Optional)_ svelte action parameters |
-
-**Returns:**
-
-ActionReturn&lt;Partial&lt;[CopyParameters](./copy.copyparameters.md)<!-- -->&lt;K&gt;&gt;, [CopyAttributes](./copy.copyattributes.md)<!-- -->&gt;
-
-svelte 
-
