--- packages/actions/copy/api/docs/copy.copyattributes._on_copied_.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@svelte-put/copy](./copy.md) &gt; [CopyAttributes](./copy.copyattributes.md) &gt; ["on:copied"](./copy.copyattributes._on_copied_.md)
-
-## CopyAttributes."on:copied" property
-
-**Signature:**
-
-```typescript
-'on:copied'?: (event: CustomEvent<CopyDetail>) => void;
-```
