--- packages/actions/copy/api/docs/copy.copyattributes.md
@@ -1,37 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@svelte-put/copy](./copy.md) &gt; [CopyAttributes](./copy.copyattributes.md)
-
-## CopyAttributes interface
-
-Additional attributes extended from `svelte-put/copy`
-
-**Signature:**
-
-```typescript
-export interface CopyAttributes 
-```
-
-## Remarks
-
-The ambient types for these extended attributes should be available automatically whenever `svelte-put/copy` is imported.
-
-```html
-<script lang="ts">
-  import { copy } from '@svelte-put/copy';
-  let trigger: HTMLElement;
-</script>
-
-<!-- on:copied should be typed -->
-<div
-  use:copy={{ trigger }}
-  on:copied
-/>
-```
-
-## Properties
-
-|  Property | Modifiers | Type | Description |
-|  --- | --- | --- | --- |
-|  ["on:copied"?](./copy.copyattributes._on_copied_.md) |  | (event: CustomEvent&lt;[CopyDetail](./copy.copydetail.md)<!-- -->&gt;) =&gt; void | _(Optional)_ |
-
