--- packages/actions/copy/api/docs/copy.copydetail.md
@@ -1,20 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@svelte-put/copy](./copy.md) &gt; [CopyDetail](./copy.copydetail.md)
-
-## CopyDetail interface
-
-`detail` payload for `copy` `CustomEvent`
-
-**Signature:**
-
-```typescript
-export interface CopyDetail 
-```
-
-## Properties
-
-|  Property | Modifiers | Type | Description |
-|  --- | --- | --- | --- |
-|  [text](./copy.copydetail.text.md) |  | string | the copied text |
-
