--- packages/actions/copy/api/docs/copy.copydetail.text.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@svelte-put/copy](./copy.md) &gt; [CopyDetail](./copy.copydetail.md) &gt; [text](./copy.copydetail.text.md)
-
-## CopyDetail.text property
-
-the copied text
-
-**Signature:**
-
-```typescript
-text: string;
-```
