--- packages/actions/copy/api/docs/copy.copyparameters.enabled.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@svelte-put/copy](./copy.md) &gt; [CopyParameters](./copy.copyparameters.md) &gt; [enabled](./copy.copyparameters.enabled.md)
-
-## CopyParameters.enabled property
-
-whether to activate the action. Default to `true`
-
-**Signature:**
-
-```typescript
-enabled: boolean;
-```
