--- packages/actions/copy/api/docs/copy.copyparameters.event.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@svelte-put/copy](./copy.md) &gt; [CopyParameters](./copy.copyparameters.md) &gt; [event](./copy.copyparameters.event.md)
-
-## CopyParameters.event property
-
-The event to trigger the copy action, passed to `addEventListener` on `trigger`<!-- -->. Default to `click`
-
-**Signature:**
-
-```typescript
-event: K | K[];
-```
