--- packages/actions/copy/api/docs/copy.copyparameters.synthetic.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@svelte-put/copy](./copy.md) &gt; [CopyParameters](./copy.copyparameters.md) &gt; [synthetic](./copy.copyparameters.synthetic.md)
-
-## CopyParameters.synthetic property
-
-whether to dispatch a synthetic `copy` event. Defaults to `false`
-
-**Signature:**
-
-```typescript
-synthetic: boolean;
-```
