--- packages/actions/copy/api/docs/copy.copyparameters.text.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@svelte-put/copy](./copy.md) &gt; [CopyParameters](./copy.copyparameters.md) &gt; [text](./copy.copyparameters.text.md)
-
-## CopyParameters.text property
-
-the text to copy, or a sync/async function that returns it
-
-**Signature:**
-
-```typescript
-text: string | TextResolver<K>;
-```
