--- packages/actions/copy/api/docs/copy.copyparameters.trigger.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@svelte-put/copy](./copy.md) &gt; [CopyParameters](./copy.copyparameters.md) &gt; [trigger](./copy.copyparameters.trigger.md)
-
-## CopyParameters.trigger property
-
-the `HTMLElement` to register event on. Default to the `node` on which the action is registered.
-
-**Signature:**
-
-```typescript
-trigger: HTMLElement;
-```
