--- packages/actions/copy/api/docs/copy.copytoclipboard.md
@@ -1,28 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@svelte-put/copy](./copy.md) &gt; [copyToClipboard](./copy.copytoclipboard.md)
-
-## copyToClipboard() function
-
-Copy a text to the clipboard. Will attempt to use the new `navigator.clipboard` API, but will fallback to the `execCommand()` if `navigator.clipboard` is not available.
-
-**Signature:**
-
-```typescript
-export declare function copyToClipboard(text: string): Promise<void> | undefined;
-```
-
-## Parameters
-
-|  Parameter | Type | Description |
-|  --- | --- | --- |
-|  text | string | text to copy |
-
-**Returns:**
-
-Promise&lt;void&gt; \| undefined
-
-## Remarks
-
-This is shamelessly "copied" from [vitepress source code](https://github.com/vuejs/vitepress/blob/43c89d66c0d8c87e244a0a0e73a897509b977e65/src/client/app/composables/copyCode.ts)
-
