--- packages/actions/copy/api/docs/copy.md
@@ -1,30 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@svelte-put/copy](./copy.md)
-
-## copy package
-
-svelte action and utilities for copying text to clipboard
-
-## Functions
-
-|  Function | Description |
-|  --- | --- |
-|  [copy(node, parameters)](./copy.copy.md) | Copy text to clipboard on <code>click</code> event |
-|  [copyToClipboard(text)](./copy.copytoclipboard.md) | Copy a text to the clipboard. Will attempt to use the new <code>navigator.clipboard</code> API, but will fallback to the <code>execCommand()</code> if <code>navigator.clipboard</code> is not available. |
-
-## Interfaces
-
-|  Interface | Description |
-|  --- | --- |
-|  [CopyAttributes](./copy.copyattributes.md) | Additional attributes extended from <code>svelte-put/copy</code> |
-|  [CopyDetail](./copy.copydetail.md) | <code>detail</code> payload for <code>copy</code> <code>CustomEvent</code> |
-|  [CopyParameters](./copy.copyparameters.md) | svelte action parameters to config behavior of <code>copy</code> |
-|  [TextResolverInput](./copy.textresolverinput.md) | the input passed to [TextResolver](./copy.textresolver.md) |
-
-## Type Aliases
-
-|  Type Alias | Description |
-|  --- | --- |
-|  [TextResolver](./copy.textresolver.md) | a sync/async function that return the text to be copied |
-
