--- packages/actions/copy/api/docs/copy.textresolver.md
@@ -1,15 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@svelte-put/copy](./copy.md) &gt; [TextResolver](./copy.textresolver.md)
-
-## TextResolver type
-
-a sync/async function that return the text to be copied
-
-**Signature:**
-
-```typescript
-export type TextResolver<K extends keyof HTMLElementEventMap> = (input: TextResolverInput<K>) => string | Promise<string>;
-```
-**References:** [TextResolverInput](./copy.textresolverinput.md)
-
