--- packages/actions/copy/api/docs/copy.textresolverinput.event.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@svelte-put/copy](./copy.md) &gt; [TextResolverInput](./copy.textresolverinput.md) &gt; [event](./copy.textresolverinput.event.md)
-
-## TextResolverInput.event property
-
-the event that triggered the listener
-
-**Signature:**
-
-```typescript
-event: HTMLElementEventMap[K];
-```
