--- packages/actions/copy/api/docs/copy.textresolverinput.node.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@svelte-put/copy](./copy.md) &gt; [TextResolverInput](./copy.textresolverinput.md) &gt; [node](./copy.textresolverinput.node.md)
-
-## TextResolverInput.node property
-
-the node the action was registered on
-
-**Signature:**
-
-```typescript
-node: HTMLElement;
-```
