--- packages/actions/copy/api/docs/copy.textresolverinput.trigger.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@svelte-put/copy](./copy.md) &gt; [TextResolverInput](./copy.textresolverinput.md) &gt; [trigger](./copy.textresolverinput.trigger.md)
-
-## TextResolverInput.trigger property
-
-the trigger element, as specified by the action parameter, otherwise same as `node`
-
-**Signature:**
-
-```typescript
-trigger: HTMLElement;
-```
