--- packages/actions/copy/api/docs/index.md
@@ -1,12 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md)
-
-## API Reference
-
-## Packages
-
-|  Package | Description |
-|  --- | --- |
-|  [@svelte-put/copy](./copy.md) | svelte action and utilities for copying text to clipboard |
-
