+++ packages/actions/copy/package.json
@@ -2,23 +2,21 @@
   "name": "@svelte-put/copy",
   "version": "2.0.0",
   "description": "Svelte action & utilities for copying text to clipboard",
-  "main": "dist/index.js",
-  "module": "dist/index.js",
-  "types": "dist/src/index.d.ts",
+  "main": "src/index.js",
+  "module": "src/index.js",
+  "types": "types/index.d.ts",
   "type": "module",
   "publishConfig": {
     "access": "public"
   },
   "files": [
-    "dist"
+    "src"
   ],
   "scripts": {
     "lint": "eslint --ignore-path .gitignore \"./**/*/*{ts,js,cjs}\"",
     "format": "prettier --check --ignore-path .gitignore --plugin-search-dir=. \"./**/*.{ts,js,cjs}\"",
-    "build": "rollup -c --configPlugin typescript",
-    "api:extract": "api-extractor run --local --verbose",
-    "api:document": "api-documenter markdown -i api/temp -o api/docs",
-    "api": "run-s api:extract api:document"
+    "dts": "node scripts/generate-dts-bundle.js",
+    "prepublishOnly": "pnpm dts"
   },
   "repository": {
     "type": "git",
@@ -38,14 +36,7 @@
   },
   "homepage": "https://svelte-put.vnphanquang.com/docs/copy",
   "devDependencies": {
-    "@microsoft/api-documenter": "^7.22.25",
-    "@microsoft/api-extractor": "^7.36.1",
-    "@rollup/plugin-typescript": "^11.0.0",
-    "@svelte-put/apirc": "workspace:*",
     "@svelte-put/tsconfig": "workspace:*",
-    "npm-run-all": "^4.1.5",
-    "rollup": "^3.20.2",
-    "tslib": "^2.5.0",
-    "typescript": "^5.0.3"
+    "dts-buddy": "^0.1.7"
   }
 }
