+++ packages/actions/copy/src/copy.helpers.js
@@ -7,9 +7,9 @@
  *
  * This is shamelessly "copied" from {@link https://github.com/vuejs/vitepress/blob/43c89d66c0d8c87e244a0a0e73a897509b977e65/src/client/app/composables/copyCode.ts | vitepress source code}
  *
- * @param text - text to copy
+ * @param {string} text - text to copy
  */
-export function copyToClipboard(text: string) {
+export function copyToClipboard(text) {
   try {
     return navigator.clipboard.writeText(text);
   } catch {
@@ -46,7 +46,7 @@ export function copyToClipboard(text: string) {
 
     // Get the focus back on the previously focused element, if any
     if (previouslyFocusedElement) {
-      (previouslyFocusedElement as HTMLElement).focus();
+      /** @type {HTMLElement} */ (previouslyFocusedElement).focus();
     }
   }
 }
