+++ packages/actions/copy/tsconfig.json
@@ -1,8 +1,4 @@
 {
-  "include": ["src/**/*.ts", "rollup.config.ts"],
-  "extends": "@svelte-put/tsconfig/base.package.json",
-  "compilerOptions": {
-    "outDir": "./dist",
-    "declarationDir": "./dist"
-  }
+  "include": ["src/**/*"],
+  "extends": "@svelte-put/tsconfig/base.package.js.json"
 }
